refactor: restructure directories for code editors and viewers
