+++ src/components/Canary/CanaryPopup/CanaryCheckDetailsSpec.tsx
@@ -1,6 +1,6 @@
+import { HealthCheck } from "@flanksource-ui/api/types/health";
+import { JSONViewer } from "@flanksource-ui/ui/Code/JSONViewer";
 import { useMemo } from "react";
-import { HealthCheck } from "../../../api/types/health";
-import { JSONViewer } from "../../../ui/JSONViewer";
 
 type CanaryCheckDetailsProps = {
   check: Pick<Partial<HealthCheck>, "canary_id" | "spec">;
